+++ NEWS.md
@@ -1,5 +1,6 @@
 # automerge (development version)
 
+* Adds `am_close()` for explicitly closing a document.
 * Updates MSRV to 1.84.
 
 # automerge 0.2.0